CubeTextureをファイル分割した
