+++ src/js/resource/cube-texture/ids.ts
@@ -1,0 +1,4 @@
+/** キューブテクスチャIDをまとめたもの */
+export const CUBE_TEXTURE_IDS = {
+  BlueSky: "BlueSky",
+};
