+++ src/js/resource/loading/full-resource-configs.ts
@@ -1,5 +1,5 @@
 import { CANVAS_IMAGE_CONFIGS } from "../canvas-image";
-import { CUBE_TEXTURE_CONFIGS } from "../cube-texture";
+import { CUBE_TEXTURE_CONFIGS } from "../cube-texture/configs";
 import { GLTF_CONFIGS } from "../gltf/configs";
 import { SOUND_CONFIGS } from "../sound/configs";
 import {
